Use `int` instead of `size_t`.

+++ week_02/matrix_multiplication/matrix/helpers.h
@@ -1,7 +1,7 @@
 #include "../common/headers.h"
 #include "matrix.h"
 
-size_t n = 1000;
+int n = 1000;
 
 Matrix mtx_a;
 Matrix mtx_b;
